Remove unnecessary container reducer
